made signup, login and app pages more lean
