--- african-market-place/src/components/pages/Login/LoginContainer.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import RenderLoginPage from './RenderLoginPage';
-
-const LoginContainer = () => {
-  return (
-    <>
-      <RenderLoginPage />
-    </>
-  );
-};
-
-export default LoginContainer;
