+++ african-market-place/src/components/pages/Login/index.js
@@ -1,1 +1,70 @@
-export { default as LoginPage } from './LoginContainer';
+import React, { useState } from 'react';
+
+export const LoginPage = () => {
+  const initialForm = {
+    username: '',
+    password: '',
+  };
+
+  const [form, setForm] = useState(initialForm);
+
+  const onChange = e => {
+    const { name, value } = e.target;
+    setForm({
+      ...form,
+      [name]: value,
+    });
+  };
+
+  const onSubmit = e => {
+    e.preventDefault();
+    // redirect to home page on successful validation
+  };
+
+  return (
+    <div id="page-wrapper">
+      <div className="landing is-preload">
+        <section id="banner">
+          <form onSubmit={onSubmit}>
+            <div className="inner">
+              <h2>Login</h2>
+              <label
+                style={{
+                  maxWidth: '200px',
+                  margin: '40px auto',
+                }}
+              >
+                username
+                <input
+                  type="text"
+                  name="username"
+                  value={form.username}
+                  onChange={onChange}
+                />
+              </label>
+              <label
+                style={{
+                  maxWidth: '200px',
+                  margin: '40px auto',
+                }}
+              >
+                password
+                <input
+                  type="password"
+                  name="password"
+                  value={form.password}
+                  onChange={onChange}
+                />
+              </label>
+              <ul className="actions special">
+                <li>
+                  <button className="button primary">Login</button>
+                </li>
+              </ul>
+            </div>
+          </form>
+        </section>
+      </div>
+    </div>
+  );
+};
