--- african-market-place/src/components/pages/SignUp/SignUpContainer.js
@@ -1,9 +1,0 @@
-import RenderSignUpPage from './RenderSignUpPage';
-
-const SignUpContainer = () => (
-  <>
-    <RenderSignUpPage />
-  </>
-);
-
-export default SignUpContainer;
